fix bugs

+++ mapper/route.go
@@ -14,7 +14,7 @@ type Route struct {
 	AuthorizeTypeKey string // 区分单例登录的Key
 	CustomToken      int    // 是否自定义授权token
 	Description      string
-	Sort             int
+	Sort             int64
 	CreateTime       time.Time `xorm:"created"`
 	UpdateTime       time.Time `xorm:"updated"`
 }
